+++ Easy/Problem-344.cpp
@@ -10,7 +10,7 @@
  * You may assume all the characters consist of printable ascii characters.
  */
 void reverseString(std::vector<char>& s) {
-    for(int i = 0, j = s.size() - 1; i < j; ++i, --j)
+    for(size_t i = 0, j = s.size() - 1; i < j; ++i, --j)
         std::swap(s[i], s[j]);
 }
